Incomplete: Refactor for adding roles to users and assigned_users to projects

+++ server/controllers/project.ts
@@ -1,7 +1,6 @@
 import { Request, Response } from 'express';
 import { getRepository } from 'typeorm';
 import { Project } from '../entity/Project';
-import { User } from '../entity/User';
 import logger from '../middleware/logger';
 
 export const createProject = async (request: Request, response: Response) => {
@@ -11,18 +10,16 @@ export const createProject = async (request: Request, response: Response) => {
     project.title = request.body.title;
     project.description = request.body.description;
 
-    const userRepository = getRepository(User);
-    const user = await userRepository.findOne(request.session.userId);
-    if (!user) {
-      response.status(401).send();
+    if (response.locals.user) {
+      project.user = response.locals.user;
+      project.assigned_users = [response.locals.user];
+      await projectRepository.save(project);
+      logger.info('Project created successfully: ' + project.id);
+      response.status(201).send({ field: 'alert', message: 'Project successfully created.', project: project });
+    } else {
+      response.status(401).send({ message: 'Access denied' });
       return;
     }
-    project.user = user;
-
-    await projectRepository.save(project);
-    logger.info('Project created successfully: ' + project.id);
-    const _project = { ...project, user: undefined };
-    response.status(201).send({ field: 'alert', message: 'Project successfully created.', project: _project });
   } catch (error) {
     logger.error(error);
     response.status(500).send({ message: error.message });
@@ -31,8 +28,12 @@ export const createProject = async (request: Request, response: Response) => {
 
 export const updateProject = async (request: Request, response: Response) => {
   try {
-    const updatedProject = await response.locals.projectRepository.save({ id: response.locals.project.id, ...request.body });
-    response.status(200).send({ field: 'alert', message: 'Project successfully updated.', project: updatedProject });
+    if (response.locals.projectRepository && response.locals.project) {
+      const updatedProject = await response.locals.projectRepository.save({ id: response.locals.project.id, ...request.body });
+      response.status(200).send({ field: 'alert', message: 'Project successfully updated.', project: updatedProject });
+    } else {
+      response.status(401).send({ field: 'alert', message: 'Access denied' });
+    }
   } catch (error) {
     logger.error(error);
     response.status(500).send({ message: error.message });
@@ -41,8 +42,12 @@ export const updateProject = async (request: Request, response: Response) => {
 
 export const deleteProject = async (_: Request, response: Response) => {
   try {
-    await response.locals.projectRepository.remove(response.locals.project);
-    response.status(200).send({ field: 'alert', message: 'Project successfully deleted.' });
+    if (response.locals.projectRepository && response.locals.project) {
+      await response.locals.projectRepository.remove(response.locals.project);
+      response.status(200).send({ field: 'alert', message: 'Project successfully deleted.' });
+    } else {
+      response.status(401).send({ field: 'alert', message: 'Access denied' });
+    }
   } catch (error) {
     logger.error(error);
     response.status(500).send({ message: error.message });
@@ -51,10 +56,8 @@ export const deleteProject = async (_: Request, response: Response) => {
 
 export const getProjects = async (request: Request, response: Response) => {
   try {
-    const userRepository = getRepository(User);
-    const user = await userRepository.findOne(request.session.userId);
     const projectRepository = getRepository(Project);
-    const projects = await projectRepository.find({ user });
+    const projects: Project[] = await projectRepository.createQueryBuilder('project').leftJoinAndSelect('project.assigned_users', 'user').where(`user.id = ${request.session.userId}`).getMany();
 
     response.status(200).send({ projects });
   } catch (error) {
@@ -62,18 +65,21 @@ export const getProjects = async (request: Request, response: Response) => {
   }
 };
 
-export const getProject = async (request: Request, response: Response) => {
+export const getProject = async (_: Request, response: Response) => {
   try {
-    const projectRepository = getRepository(Project);
-    const project = await projectRepository.findOne(request.params.id);
-
-    if (!project) {
+    if (!response.locals.project) {
       response.status(404).send({ message: 'Project not found' });
       return;
     }
 
-    response.status(200).send({ project });
+    response.status(200).send({ project: response.locals.project });
   } catch (error) {
     response.status(500).send({ error: error.message });
   }
 };
+
+// TODO
+
+// export const addUserToProject = async (request: Request, response: Response) => {};
+
+// export const removeUserFromProject = async (request: Request, response: Response) => {};
